% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{model_fit}
\alias{model_fit}
\title{Model Fit results}
\usage{
model_fit(Results, INP, outdir)
}
\arguments{
\item{Results}{TMB.LBSPR Results}

\item{INP}{Input Data}

\item{outdir}{Location to store outputs to file. Defaults to TMB.LBSPR directory in the system's Home.}
}
\description{
Takes the model residuals from run_analyses and generates diagnostic plots
}
