% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_outputs.R
\name{process_outputs}
\alias{process_outputs}
\title{TMB.LPSBR model output}
\usage{
process_outputs(D, TYPE = c("Both", "Survey only", "Catch only"),
  SHOW.LC = TRUE, outdir = file.path(home = Sys.getenv("HOME"),
  "TMB.LBSPR"))
}
\arguments{
\item{D}{TMB.LBSPR MCMC dataset}

\item{TYPE}{ype of data available ("Survey only", "Catch only", "Both).}

\item{SHOW.LC}{Option to show minimum size (Lc) management results}

\item{outdir}{Location to store outputs to file. Defaults to TMB.LBSPR directory in the system's Home.}
}
\description{
Takes the output from run_analyses, a Monte Carlo dataset, and creates tables and graphs
}
