% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analyses.R
\name{run_analyses}
\alias{run_analyses}
\title{Run TMB-LBSPR Analysis}
\usage{
run_analyses(D, Species, n_iteration, n_GTG, starting, ManageF30 = TRUE,
  ManageLc30 = TRUE, NumCores = -999, Seed = 1)
}
\arguments{
\item{D}{Input Data}

\item{Species}{Species name}

\item{n_iteration}{Number of Iterations}

\item{n_GTG}{Number of growth-type groups used in simulation (default at 20)}

\item{starting}{List of starting values for the 3 estimated parameters}

\item{ManageF30}{Option to return max fishing mortality (F30) management recommendation}

\item{ManageLc30}{Option to return minimum size (Lc30) management recommendation.}

\item{NumCores}{Number of CPU cores to use (Default?)}

\item{Seed}{Psuedo-Random Number Seed}
}
\description{
Runs TMB LBSPR model and calls process_results function.
}
